Extract middleware behavior to external file.
